+++ basic/models/AircraftConfiguration.php
@@ -14,6 +14,7 @@ use Yii;
  * @property int $cargo_capacity
  * @property int $crew
  * @property int $mtow
+ * @property int $bew
  *
  * @property AircraftType $aircraftType
  * @property Aircraft[] $aircrafts
@@ -34,9 +35,9 @@ class AircraftConfiguration extends \yii\db\ActiveRecord
     public function rules()
     {
         return [
-            [['aircraft_type_id', 'name', 'pax_capacity', 'cargo_capacity', 'crew', 'mtow'], 'required'],
+            [['aircraft_type_id', 'name', 'pax_capacity', 'cargo_capacity', 'crew', 'mtow', 'bew'], 'required'],
             [['aircraft_type_id', 'pax_capacity', 'cargo_capacity'], 'integer', 'min' => 0],
-            [['crew', 'mtow'], 'integer', 'min' => 1],
+            [['crew', 'mtow', 'bew'], 'integer', 'min' => 1],
             [['name'], 'string', 'max' => 20],
             [['name'], 'trim'],
             [['name'], 'unique', 'targetAttribute' => ['aircraft_type_id', 'name']],
@@ -57,6 +58,7 @@ class AircraftConfiguration extends \yii\db\ActiveRecord
             'cargo_capacity' => Yii::t('app', 'Cargo Capacity'),
             'crew' => Yii::t('app', 'Crew'),
             'mtow' => Yii::t('app', 'MTOW (Kg)'),
+            'bew' => Yii::t('app', 'BEW (Kg)'),
         ];
     }