+++ basic/tests/unit/models/AircraftTest.php
@@ -45,6 +45,7 @@ class AircraftTest extends BaseUnitTest
             'cargo_capacity' => 2000,
             'crew' => 5,
             'mtow' => 79016,
+            'bew' => 41413,
         ]);
         $this->standardConfig->save();