+++ basic/tests/unit/models/SubmittedFlightPlanTest.php
@@ -159,6 +159,7 @@ class SubmittedFlightPlanTest extends BaseUnitTest
             'cargo_capacity' => 2000,
             'crew' => 5,
             'mtow' => 79016,
+            'bew' => 41413,
         ]);
         $config->save();