+++ basic/views/aircraft-configuration/_form.php
@@ -27,6 +27,8 @@ use yii\widgets\ActiveForm;
 
     <?= $form->field($model, 'mtow')->textInput() ?>
 
+    <?= $form->field($model, 'bew')->textInput() ?>
+
     <div class="form-group">
         <?= Html::submitButton(Yii::t('app', 'Save'), ['class' => 'btn btn-success']) ?>
     </div>