+++ basic/views/aircraft-configuration/view.php
@@ -45,6 +45,7 @@ $this->params['breadcrumbs'][] = $this->title;
             'cargo_capacity',
             'crew',
             'mtow',
+            'bew',
         ],
     ]) ?>